[SPARK-47399][SQL] Disable generated columns on expressions with collations
### What changes were proposed in this pull request?
Disable the ability to use collations in expressions for generated columns.

### Why are the changes needed?
Changing the collation of a column or even just changing the ICU version could lead to a differences in the resulting expression so it would be best if we simply disable it for now.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

With new unit tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45520 from stefankandic/disableGeneratedColumnsCollation.

Authored-by: Stefan Kandic <stefan.kandic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>