[SPARK-47423][SQL] Collations - Set operation support for strings with collations
### What changes were proposed in this pull request?

This PR fixes support for set operations for strings with collations different from `UTF8_BINARY`. The fix is not strictly related to set operations and may resolve other problems in collation space. The fix is to add default value for `StringType` with collation. Previously the matching pattern would not catch the `StringType` with collation case and fix is simply to do pattern matching on `st: StringType` instead of relying on `StringType` match.

### Why are the changes needed?

Fixing behaviour of set operations.

### Does this PR introduce _any_ user-facing change?

Yes - fixing the logic that previously didn't work.

### How was this patch tested?

Golden file tests are added.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45536 from dbatomic/collations_and_set_ops.

Authored-by: Aleksandar Tomic <aleksandar.tomic@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>