[SPARK-47380][CONNECT] Ensure on the server side that the SparkSession is the same
### What changes were proposed in this pull request?

In this PR we change the client behaviour to send the previously observed server session id so that the server can validate that the client used to talk with this specific session. Previously this was only validated on the client side which made the server actually execute the request for the wrong session before throwing on the client side (once the response from the server was obtained).

### Why are the changes needed?
The server can execute the client command on the wrong spark session before client figuring out it's the different session.

### Does this PR introduce _any_ user-facing change?
The error message now pops up differently (it used to be a slightly different message when validated on the client).

### How was this patch tested?
Existing unit tests, add new unit test, e2e test added, manual testing

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #45499 from nemanja-boric-databricks/workspace.

Authored-by: Nemanja Boric <nemanja.boric@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>