[SPARK-47398][SQL] Extract a trait for InMemoryTableScanExec to allow for extending functionality
### What changes were proposed in this pull request?
We are proposing to allow the users to have a custom `InMemoryTableScanExec`. To accomplish this we can follow the same convention we followed for `ShuffleExchangeLike` and `BroadcastExchangeLike`

### Why are the changes needed?
In the PR added by ulysses-you, we are wrapping `InMemoryTableScanExec` inside `TableCacheQueryStageExec`. This could potentially cause problems, especially in the RAPIDS Accelerator for Apache Spark, where we replace `InMemoryTableScanExec` with a customized version that has optimizations needed by us. This situation could lead to the loss of benefits from [SPARK-42101](https://issues.apache.org/jira/browse/SPARK-42101) or even result in Spark throwing an Exception.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Ran the existing tests

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #45525 from razajafri/extract-inmem-trait.

Authored-by: Raza Jafri <rjafri@nvidia.com>
Signed-off-by: Thomas Graves <tgraves@apache.org>