[MINOR][CORE] When failed to canceling the job group, add a warning log
### What changes were proposed in this pull request?
The pr aims to add add a warning log when failed to canceling the job group.

### Why are the changes needed?
In production, we found that when there are many jobs, if we specify `a just-submitted job group` and `cancel` it using the API `SparkContext#cancelJobGroup(groupId: String)`, the cancellation will fail (`silently`) because the job has not yet entered `activeJobs`, and there is `no warning` prompt.
(PS: Due to historical reasons, we are not currently using the newer API `SparkContext#cancelJobGroupAndFutureJobs(groupId: String)`)

I propose to add `a warning log` here to help troubleshoot `issues`.

### Does this PR introduce _any_ user-facing change?
Yes, only for log.

### How was this patch tested?
N/A

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #45722 from panbingkun/handleJobGroupCancelled_add_log.

Authored-by: panbingkun <panbingkun@baidu.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>