[SPARK-42040][SQL] SPJ: Introduce a new API for V2 input partition to report partition statistics
### What changes were proposed in this pull request?
This PR introduces a new V2 min-in HasPartitionStatistics which can be used to return the partition statistics of a InputPartition.

### Why are the changes needed?
As part of the Storage Partitioned Join work ([SPIP](https://issues.apache.org/jira/browse/SPARK-37166)), we'll need to introduce a way for a V2 InputPartition to return its partition size,  It's useful for a InputPartition to also report its size (in bytes), so that Spark can use the info to decide whether partition grouping should be applied or not.

This will be used later in follow-up PRs.

### Does this PR introduce _any_ user-facing change?
Yes, a new V2 mix-in HasPartitionStatistics will be introduced.

### How was this patch tested?
Extended InMemoryTable to support this new interface, and added a new unit test to verify the API, and the mock the partition statistics to test.

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #45314 from zhuqi-lucas/SPARK-42040.

Authored-by: qzhu <qzhu@cloudera.com>
Signed-off-by: Your Name <chao@openai.com>