[SPARK-47616][SQL] Add User Document for Mapping Spark SQL Data Types from MySQL
### What changes were proposed in this pull request?

This PR added a User Document for Mapping Spark SQL Data Types from MySQL. The write side document is not included yet which might need further verification.

### Why are the changes needed?

Now, the conversion of data types from MySQL to Spark SQL is solid. End users can refer to it.

It's also time for maintainers to have an overall perspective to review the changes in the coming 4.0.0

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?

add some test for missing MySQL data types

![image](https://github.com/apache/spark/assets/8326978/da8eb241-1d05-4a01-b038-d7d329193674)

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #45736 from yaooqinn/SPARK-47616.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>