[SPARK-47705][INFRA] Sort LogKey alphabetically and build a test to ensure it
### What changes were proposed in this pull request?

This PR adds a unit test to ensure that the fields of the `LogKey` enumeration are sorted alphabetically, as specified by https://issues.apache.org/jira/browse/SPARK-47705.

### Why are the changes needed?

This will make sure that the fields of the enumeration remain easy to read in the future as we add more cases.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

This PR adds testing coverage only.

### Was this patch authored or co-authored using generative AI tooling?

GitHub copilot offered some suggestions, but I rejected them

Closes #45857 from dtenedor/logs.

Authored-by: Daniel Tenedorio <daniel.tenedorio@databricks.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>