[SPARK-47708][CONNECT] Do not log gRPC exception to stderr in PySpark
### What changes were proposed in this pull request?

Currently if there's any gRPC exception, instead of just handling it, the PySpark's gRPC error handler is going to print it out to the stderr, not allowing the user to cleanly ignore the exception by using try/except control flow statement.

In this PR we are removing the logger.exception call and we rely on the downstream exception mechanism to report this to the user.

### Why are the changes needed?

Without this change, there's no way that the user ignores the gRPC error without piping the stderr to /dev/null or equivalent.

### Does this PR introduce _any_ user-facing change?

Yes, the stderr will not have the exception trace written twice.

### How was this patch tested?

Existing tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45840 from nemanja-boric-databricks/no-log.

Authored-by: Nemanja Boric <nemanja.boric@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>