[SPARK-47719][SQL] Change spark.sql.legacy.timeParserPolicy default to CORRECTED
### What changes were proposed in this pull request?

We changed the time parser policy in Spark 3.0.0.
The config has since defaulted to raise an exception if there is a potential conflict between teh legacy and the new policy.
Spark 4.0.0 is a good time to default to the new policy

### Why are the changes needed?

Move the product forward and retire legacy behavior over time.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Run existing unit tests and verify changes.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #45859 from srielau/SPARK-47719-parser-policy-default-to-corrected.

Lead-authored-by: Serge Rielau <serge@rielau.com>
Co-authored-by: Wenchen Fan <cloud0fan@gmail.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>