[SPARK-47727][PYTHON] Make SparkConf to root level to for both SparkSession and SparkContext
### What changes were proposed in this pull request?

This PR proposes to make SparkConf to root level to for both `SparkSession` and `SparkContext`.

### Why are the changes needed?

`SparkConf` is special. `SparkSession.builder.options` can take it as an option, and this instance can be created without JVM access. So it can be shared with pure Python `pysaprk-connect` package.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

CI in this PR should verify them.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45873 from HyukjinKwon/SPARK-47727.

Lead-authored-by: Hyukjin Kwon <gurwls223@apache.org>
Co-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>