[SPARK-47736][SQL] Add support for AbstractArrayType
### What changes were proposed in this pull request?
Addition of abstract arraytype which accepts StringTypeCollated as elementType. Changes in this PR https://github.com/apache/spark/pull/45693 work for ArrayJoin, but will not work in general for other functions. This PR introduces a change to give an interface for all functions. Merge only after #45693.

### Why are the changes needed?
This is needed in order to enable functions to use collated arrays.

### Does this PR introduce _any_ user-facing change?
Yes, collation functions will work.

### How was this patch tested?
Test for array_join added to `CollationSuite`

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #45891 from mihailom-db/SPARK-47736.

Authored-by: Mihailo Milosevic <mihailo.milosevic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>