[SPARK-47813][SQL] Replace getArrayDimension with updateExtraColumnMeta
### What changes were proposed in this pull request?

[SPARK-47754](https://issues.apache.org/jira/browse/SPARK-47754) introduced a new developer API called `getArrayDimension`.

This PR expands the scope of `getArrayDimension` and renames it to `updateExtraColumnMeta`. Just as their names said, `getArrayDimension` handles only one column of metadata that is the dimension of an array, while `updateExtraColumnMeta` can retrieve any type of metadata based on the given ResultSetMetadata and Connection. This is much more general and useful and reduces the number of potentially new Developer APIs in the same shape. Also the current parameters for
getArrayDimension might not be enough for other dialects

### Why are the changes needed?

Refactoring unreleased Developer APIs to make it more sustainable

### Does this PR introduce _any_ user-facing change?

no, unreleased API change

### How was this patch tested?

existing ut

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #46006 from yaooqinn/SPARK-47813.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Kent Yao <yao@apache.org>