[SPARK-47784][SS] Merge TTLMode and TimeoutMode into a single TimeMode
### What changes were proposed in this pull request?

This PR merges the `TimeoutMode` and `TTLMode` parameter for `transformWithState` into a single `TimeMode`. Currently, users need to specify the notion of time (ProcessingTime/EventTime) for timers and ttl separately. This allows users to use a single parameter.

We do not expect users to use mix/match EventTime/ProcessingTime for timers and ttl in a single query because it makes hard to reason about the time semantics (when will timer be fired?, when will the state be evicted? etc.). Its simpler to stick to one notion of time throughout timers and ttl.

### Why are the changes needed?

Changes are needed to simplify Arbitrary State API `transformWithState` interface by merging TTLMode/TimeoutMode into a single TimeMode.

### Does this PR introduce _any_ user-facing change?

Yes, this PR changes the API parameters for `transformWithState`.

### How was this patch tested?

All existing testcases for `transformWithState` API pass.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #45960 from sahnib/introduce-timeMode.

Authored-by: Bhuwan Sahni <bhuwan.sahni@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>