[SPARK-47812][CONNECT] Support Serialization of SparkSession for ForEachBatch worker
### What changes were proposed in this pull request?

This patch adds support to register custom dispatch handlers when serializing objects using the provided Cloudpickle library. This is necessary to provide compatibility when executing ForEachBatch functions in structured streaming.

A typical example for this behavior is the following test case:

```python
def curried_function(df):
    def inner(batch_df, batch_id):
        df.createOrReplaceTempView("updates")
        batch_df.createOrReplaceTempView("batch_updates")

    return inner

df = spark.readStream.format("text").load("python/test_support/sql/streaming")
other_df = self.spark.range(100)
df.writeStream.foreachBatch(curried_function(other_df)).start()
```
Here we curry a DataFrame into the function called during ForEachBatch and effectively passing state. Until now, serializing DataFrames and SparkSessions in Spark Connect was not possible since the SparkSession carries the open GPRC connection and the DataFrame itself overrides certain magic methods that make pickling fail.

To make serializing Spark Sessions possible, we register a custom session constructor, that simply returns the current active session, during the serialization of the ForEachBatch function. Now, when the ForEachBatch worker starts the execution it already creates and registers an active SparkSession. To serialize and reconstruct the DataFrame we simply have to pass in the session and the plan, the remaining attributes do not carry a permanent state.

To avoid modifying any global behavior, the serialization handlers are not registered for all cases but only when the ForEachBatch and ForEach handlers are called. This is to make sure that we don't unexpectedly change behavior.

### Why are the changes needed?
Compatibility and Ease of Use

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Added and updated tests

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #46002 from grundprinzip/SPARK-47812.

Lead-authored-by: Martin Grund <martin.grund@databricks.com>
Co-authored-by: Martin Grund <grundprinzip@gmail.com>
Co-authored-by: Hyukjin Kwon <gurwls223@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>