[SPARK-47673][SS] Implementing TTL for ListState
### What changes were proposed in this pull request?

This PR adds support for expiring state based on TTL for ListState. Using this functionality, Spark users can specify a TTL Mode for transformWithState operator, and provide a ttlDuration for each value in ListState. TTL support for Map State will be added in future PRs. Once the ttlDuration has expired, the value will not be returned as part of get() and would be cleaned up at the end of the micro-batch.

### Why are the changes needed?

These changes are needed to support TTL for ListState. The PR supports specifying ttl for processing time.

### Does this PR introduce _any_ user-facing change?

Yes, modifies the ListState interface for specifying ttlDuration

### How was this patch tested?

Added the TransformWithListStateTTLSuite, ListStateSuite, StatefulProcessorHandleSuite
### Was this patch authored or co-authored using generative AI tooling?

No

Closes #45932 from ericm-db/ls-ttl.

Authored-by: Eric Marnadi <eric.marnadi@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>