[SPARK-47081][CONNECT][FOLLOW] Unflake Progress Execution
### Why are the changes needed?
Running some of the tests repetitively have shown that there is an edge case in which the progress reporting yields undesired behavior. This is in particular the case when the response sender loop has already sent the last expected message in the stream to indicate that the stream is finished, but the response sender would send at least one additional progress message.

This patch removes this race condition and guarantees that the `ResultComplete` message is the last message on the stream.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Added an additional UT, but existing tests cover the race condition.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #46060 from grundprinzip/SPARK-47081_3.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>