[SPARK-47877][SS][CONNECT] Speed up test_parity_listener
### What changes were proposed in this pull request?
This PR makes test_parity_listener run faster.

The test was slow because of `TestListenerSparkV1` and `TestListenerSparkV2` makes server calls and has long wait time, and the test runs on both listeners. They were created to verify the listener function with and without the new `onQueryIdle` callback.

This PR fixes the slowness by removing the V1 and V2 of that listener (now only a `TestListenerSpark`), and create lightweight `TestListenerLocalV1` and `TestListenerLocalV2` for the `onQueryIdle` verification.

### Why are the changes needed?

Faster and more stable ci

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

test only change

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #46072 from WweiL/speed-up-test.

Authored-by: Wei Liu <wei.liu@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>