[SPARK-47627][SQL] Add SQL MERGE syntax to enable schema evolution
### Why are the changes needed?

This PR introduces a syntax `WITH SCHEMA EVOLUTION` to the SQL MERGE command, which allows the user to specify automatic schema evolution for a specific operation.

```sql
MERGE WITH SCHEMA EVOLUTION
INTO tgt
USING src
ON ...
WHEN ...
```

When `WITH SCHEMA EVOLUTION` is specified, schema evolution-related features must be turned on for this single statement and only in this statement.

Spark is only responsible for recognizing the existence or absence of the syntax `WITH SCHEMA EVOLUTION`, and the result is passed down to the MERGE command. Data sources must respect the syntax and give appropriate reactions: turn on features that are categorised as "schema evolution" when the syntax does exist. For example, when the underlying table is Delta Lake, the feature "mergeSchema" will be turned on (see https://github.com/delta-io/delta/blob/c41977db3529a3139d6306abe5ded161f070982a/spark/src/main/scala/org/apache/spark/sql/delta/DeltaAnalysis.scala#L538).

### Does this PR introduce _any_ user-facing change?

Yes, see the previous section.

### How was this patch tested?

New tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45748 from xupefei/merge-schema-evolution.

Authored-by: Paddy Xu <xupaddy@gmail.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>