[SPARK-47902][SQL] Making Compute Current Time* expressions foldable
### What changes were proposed in this pull request?

This PR is a followed of [this](https://github.com/apache/spark/pull/44261) PR that made "compute current time" family of expressions `Unevaluable`. Implicitly, all `Unevaluable` expressions are also not foldable. Hence, we have a regression comparing to state prior to 44261 where "compute current time" expressions could have been used in places where constant folding is required.

Proposed change is to keep these expression `Unevaluable` in a sense that `eval`/`codeGen`  can't be called but to allow folding. This is a special case given that these expressions are supposed to be replaced by QO with literals (that are foldable) so proposal is to create new interface for this family of expressions.

### Why are the changes needed?

Explained above.

### Does this PR introduce _any_ user-facing change?

Yes.

### How was this patch tested?

Additional test is added that ensures that `CurrentDate` can be used in places that require folding.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #46120 from dbatomic/constantfoldingforcurrenttime.

Lead-authored-by: Aleksandar Tomic <aleksandar.tomic@databricks.com>
Co-authored-by: Aleksandar Tomic <150942779+dbatomic@users.noreply.github.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>